+++ runtime/kreivo/Cargo.toml
@@ -91,6 +91,7 @@ cumulus-pallet-parachain-system = {workspace = true}
 cumulus-pallet-session-benchmarking = {workspace = true}
 cumulus-pallet-xcm = {workspace = true}
 cumulus-pallet-xcmp-queue = {workspace = true}
+cumulus-primitives-aura = {workspace = true}
 cumulus-primitives-core = {workspace = true}
 cumulus-primitives-timestamp = {workspace = true}
 cumulus-primitives-utility = {workspace = true}
